+++ apps/web/src/app/customer/mypage/_components/MypageSidebar.tsx
@@ -4,12 +4,12 @@ import Link from "next/link";
 import type { ComponentType } from "react";
 import { useActivitySummary } from "../_api/use-activity-summary";
 import { MYPAGE_SIDEBAR_LINKS, type MypageSidebarLink } from "../_model/mypage-links";
-import { useLogout } from "@/shared/api/use-logout";
 import { cn } from "@/shared/lib/utils";
 import { FileText } from "@/shared/ui/icons/FileText";
 import { Home } from "@/shared/ui/icons/Home";
 import { MessageCircle } from "@/shared/ui/icons/MessageCircle";
 import { MessageSquare } from "@/shared/ui/icons/MessageSquare";
+import { LogoutButton } from "./LogoutButton";
 
 type IconComponent = ComponentType<{ className?: string }>;
 
@@ -69,21 +69,5 @@ export function MypageSidebar() {
         </Link>
       </div>
     </nav>
-  );
-}
-
-/** 로그아웃 — 세션 종료 후 로그인 화면으로 이동(#155). */
-function LogoutButton() {
-  const { mutate: logout, isPending } = useLogout();
-
-  return (
-    <button
-      type="button"
-      onClick={() => logout()}
-      disabled={isPending}
-      className="w-full rounded-button px-3 py-2.5 text-left text-[0.875rem] font-medium text-ink-3 transition hover:bg-paper hover:text-ink-2 disabled:opacity-50"
-    >
-      로그아웃
-    </button>
   );
 }
