+++ apps/web/src/app/customer/mypage/_components/mobile/CountText.tsx
@@ -1,0 +1,3 @@
+export function CountText({ value }: { value: string }) {
+  return <span className="text-[0.8125rem] font-bold text-ink-3">{value}</span>;
+}
